ISSUE=801434
git-svn-id: https://svn.baidu.com/myspace/sep/chenghuige/melt-train@296621 c4ccc3fd-b13d-49bf-a705-f618d8f4424a

+++ include/Trainers/FastRank/FastRank.h
@@ -31,6 +31,7 @@
 #include "Predictors/FastRankPredictor.h"
 #include "Prediction/Calibrate/CalibratorFactory.h"
 #include "Prediction/Instances/instances_util.h"
+#include "rabit_util.h"
 namespace gezi {
 
 	class FastRank : public ValidatingTrainer
@@ -311,7 +312,7 @@ namespace gezi {
 	protected:
 		void ConstructValidatingScoreTrackers()
 		{
-			for (int i = 0; i < _validationSets.size(); i++)
+			for (size_t i = 0; i < _validationSets.size(); i++)
 			{
 				ConstructScoreTracker(_validationSets[i], i);
 			}
